// lib/generated-plugins.ts
// Auto generated by scripts/generate-plugins.ts
import plugin0 from '@/lib/plugins/game/cek-kodam'
import plugin1 from '@/lib/plugins/tools/json-echo'
import plugin2 from '@/lib/plugins/upload/file-info'
export const plugins = [plugin0, plugin1, plugin2]
export const categories = [
  {
    "slug": "game",
    "title": "Game",
    "icon": "Boxes",
    "count": 1
  },
  {
    "slug": "tools",
    "title": "Tools",
    "icon": "Boxes",
    "count": 1
  },
  {
    "slug": "upload",
    "title": "Upload",
    "icon": "Boxes",
    "count": 1
  }
]
export function findPlugin(category:string, endpoint:string){return plugins.find(p=>p.category===category&&p.slug===endpoint)}